+++ src/solidmodels/solidmodels.jl
@@ -85,10 +85,12 @@ import DeviceLayout.Polygons:
 import DeviceLayout.Curvilinear:
     edge_type_at_vertex,
     line_arc_cornerindices,
+    arc_arc_cornerindices,
     islinear,
     round_to_curvilinearpolygon,
     rounded_corner_segment,
     rounded_corner_segment_line_arc,
+    rounded_corner_segment_arc_arc,
     to_curvilinear,
     styled_loop
 import Unitful: μm, mm, ustrip, °, uconvert, Length